+++ js/main.js
@@ -4,6 +4,7 @@ import './hashtags-pristine.js';
 import './effects.js';
 import './messages.js';
 import { loadData } from './fetch.js';
+import { showAlert } from './util.js';
 
 let photos = [];
 
@@ -13,16 +14,7 @@ const onSuccess = (data) => {
 };
 
 const onError = () => {
-  const messageAlert = document.createElement('div');
-  messageAlert.style.position = 'absolute';
-  messageAlert.style.left = 0;
-  messageAlert.style.top = 0;
-  messageAlert.style.right = 0;
-  messageAlert.style.fontSize = '30px';
-  messageAlert.style.backgroundColor = 'red';
-  messageAlert.style.textAlign = 'center';
-  messageAlert.textContent = 'Ошибка загрузки фотографий';
-  document.body.append(messageAlert);
+  showAlert();
 };
 
 loadData(onSuccess, onError);
