+++ js/util.js
@@ -19,5 +19,18 @@ const closeOnEscKeyDown = (evt, cb) => {
   }
 };
 
+const showAlert = () => {
+  const messageAlert = document.createElement('div');
+  messageAlert.style.position = 'absolute';
+  messageAlert.style.left = 0;
+  messageAlert.style.top = 0;
+  messageAlert.style.right = 0;
+  messageAlert.style.fontSize = '30px';
+  messageAlert.style.backgroundColor = 'red';
+  messageAlert.style.textAlign = 'center';
+  messageAlert.textContent = 'Ошибка загрузки фотографий';
+  document.body.append(messageAlert);
+};
+
 export {getRandomInteger, closeOnEscKeyDown};
-export{isEscapeKey};
+export{isEscapeKey, showAlert};
